-moved provider box styles + variables to teams theme
