--- packages/react/src/themes/base/components/Provider/providerBoxStyles.ts
@@ -1,9 +1,0 @@
-import { ICSSInJSStyle } from '../../../types'
-
-export default {
-  root: ({ variables: v }): ICSSInJSStyle => ({
-    background: v.background,
-    color: v.color,
-    textAlign: 'left',
-  }),
-}
